+++ src/controllers/EntryRevisionsController.php
@@ -298,14 +298,7 @@ class EntryRevisionsController extends BaseEntriesController
 
             $draft->setScenario(Element::SCENARIO_ESSENTIALS);
 
-            // todo: stop checking for propagateAll in Craft 4
-            if (
-                $draft->getIsUnpublishedDraft() &&
-                (
-                    $this->request->getBodyParam('isFresh') ||
-                    $this->request->getBodyParam('propagateAll')
-                )
-            ) {
+            if ($draft->getIsUnpublishedDraft() && $this->request->getBodyParam('isFresh')) {
                 $draft->setIsFresh();
                 $draft->propagateAll = true;
             }
@@ -490,14 +483,7 @@ class EntryRevisionsController extends BaseEntriesController
             $draft->setScenario(Element::SCENARIO_LIVE);
         }
 
-        // todo: stop checking for propagateAll in Craft 4
-        if (
-            $draft->getIsUnpublishedDraft() &&
-            (
-                $this->request->getBodyParam('isFresh') ||
-                $this->request->getBodyParam('propagateAll')
-            )
-        ) {
+        if ($draft->getIsUnpublishedDraft() && $this->request->getBodyParam('isFresh')) {
             $draft->setIsFresh();
             $draft->propagateAll = true;
         }